+++ src/kspdg/lbg1/lbg1_base.py
@@ -373,6 +373,8 @@ class LadyBanditGuardGroup1Env(Group1BaseEnv):
 
         '''
 
+        self.logger.debug("Starting get_observation ...")
+
         rf = self.vesBandit.orbit.body.non_rotating_reference_frame
 
         obs = [None] * self.PARAMS.OBSERVATION.LEN
@@ -429,6 +431,8 @@ class LadyBanditGuardGroup1Env(Group1BaseEnv):
             obs[self.PARAMS.OBSERVATION.I_GUARD_VY], \
             obs[self.PARAMS.OBSERVATION.I_GUARD_VZ]  = \
             v_g_cb__rhcbci
+        
+        self.logger.debug("get_observation complete")
 
         return np.array(obs)